Add setup file for installing the things we need
I use a bunch of applications on both my laptops and wanted a way to
install them automatically and keep them up to date.
Since I am managing my .files anyway this seemed like a good place to
do that. I wanted to just execute the resulting tangled
script but couldn't work out how to do it.
Next time I will work out how to do that so we don't need to write the
shell script to disk.