Add check to see if our dependencies are installed on MacOS
Brew has some lovely json output for it's info subcommand.
Running that through jq lets us see if a package is actually
installed.
This would also let us check versions of installed packages so we
could potentially update them if needed.
But now we have a dependency on jq so we need to make sure that is installed.