Add bootstrap to install script to install needed dependencies
The ultimate plan is to be able to use this on a bare system.
The additions added in setup.org makes some assumptions about the
state of our system that may not be true if that is the case.
We were assuming the system already had emacs installed.
On MacOS we need brew to exist so we can install emacs.
So let us make sure those are installed before we run our tangle.